requirements.txt
# Core libraries
playwright
requests
openai
openai-whisper
rich

# For audio processing
pydub
SpeechRecognition

# For environment variables
python-dotenv
httpx
websockets

piper-tts
onnxruntime
